Create obit_path.py
The orbit of a planet can be defined as its gravitationally curved path around a point in space. That point is generally a star and in some cases multiple stars. A binary star system, and multiple star systems are more common than unary star systems like our own Solar System. A binary star system centers around a foci between the two stars. For planets that are a part of a binary star system their orbits can function such as that of a planet in a unary star system if the distance between the planet and the binary star is great enough. This occurs because the gravitational force simply acts singularly upon the planet. If the planet is at a distance where the gravitational forces of both stars can act upon it  independently then the orbit of the planet may be unstable and altered drastically by one star or the other. The goal of this report is observe both of these cases and determine under what conditions can a planet sustain stable orbit in a binary star  system.